First working example
This example demonstrates how to write directly in the framebuffer of
a window, in Android.
The purpose is to show how to do pure CPU rendering in Android.

The example is not very commented, and the inclusion of the
"android_natie_app_glue" makes it a bit harder to understand what part
is essential, and what part is not.
It seems that the NDK always provide the glue and I might switch to a
CMakeLists that only uses that version instead.

Anyway, it works and the APK needed to use and compile it will be
provided afterwards.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>