Let's avoid hitting the stack every time
The window width and stride won't change during a write.
Even if the actual window change, the previously allocated buffer
properties won't.

Caching these informations tell the compiler to avoid reloading
the value again and again.
For reasons that are quite weird to me, if you don't cache these
values the compiler (clang essentially) will try to load the
values from the stack again and again...
Even though they should be read from the heap in that case...

Anyway, doing that tells the compiler to put the values in
registers only once, which can shave a few cycles per loop iteration.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>